best_subset_classifier {autoMrP}	R Documentation

Best subset classifier

Description

best_subset_classifier applies best subset classification to a data set.

Usage

best_subset_classifier(
  model,
  data.train,
  model.family,
  model.optimizer,
  n.iter,
  verbose = c(TRUE, FALSE)
)

Arguments

model
Multilevel model. A model formula describing the multilevel model to be estimated on the basis of the provided training data.

data.train
Training data. A data.frame containing the training data used to train the model.

model.family
Model family. A variable indicating the model family to be used by glmer. Defaults to binomial(link = "probit").

model.optimizer
Optimization method. A character-valued scalar describing the optimization method to be used by glmer. Defaults to "bobyqa".

n.iter
Iterations. A integer-valued scalar specifying the maximum number of function evaluations tried by the optimization method.

verbose
Verbose output. A logical vector indicating whether or not verbose output should be printed.

Examples

#not_yet

[Package autoMrP version 0.0.0.9000 Index]