gb_classifier {autoMrP}	R Documentation

GB classifier

Description

gb_classifier applies gradient boosting classification to a data set.

Usage

gb_classifier(
  form,
  distribution,
  data.train,
  n.trees,
  interaction.depth,
  n.minobsinnode,
  shrinkage,
  verbose = c(TRUE, FALSE)
)

Arguments

form
Model formula. A two-sided linear formula describing the model to be fit, with the outcome on the LHS and the covariates separated by + operators on the RHS.

distribution
Model distribution. A character string specifying the name of the distribution to be used.

data.train
Training data. A data.frame containing the training data used to train the model.

n.trees
Total number of trees. An integer-valued scalar specifying the total number of trees to be fit.

interaction.depth
Interaction depth. An integer-valued scalar specifying the maximum depth of each tree.

n.minobsinnode
Minimum number of observations in terminal nodes. An integer-valued scalar specifying the minimum number of observations in the terminal nodes of the trees.

shrinkage
Learning rate. A numeric scalar specifying the shrinkage or learning rate applied to each tree in the expansion.

verbose
Verbose output. A logical vector indicating whether or not verbose output should be printed.

Examples

#not_yet

[Package autoMrP version 0.0.0.9000 Index]