run_lasso {autoMrP}	R Documentation

Apply lasso classifier to MrP.

Description

run_lasso is a wrapper function that applies the lasso classifier to data provided by the user, evaluates prediction performance, and chooses the best-performing model.

Usage

run_lasso(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  n.iter = NULL,
  loss.unit,
  loss.fun,
  lambda,
  iterations.max,
  data,
  verbose
)

Arguments

y
Outcome variable. A character scalar containing the column name of the outcome variable in survey.

L1.x
Individual-level covariates. A character vector containing the column names of the individual-level variables in survey and census used to predict outcome y. Note that geographic unit is specified in argument L2.unit.

L2.x
Context-level covariates. A character vector containing the column names of the context-level variables in survey and census used to predict outcome y.

L2.unit
Geographic unit. A character scalar containing the column name of the geographic unit in survey and census at which outcomes should be aggregated.

L2.reg
Geographic region. A character scalar containing the column name of the geographic region in survey and census by which geographic units are grouped (L2.unit must be nested within L2.reg). Default is NULL.

n.iter
Lasso number of iterations without improvement. Either NULL or an integer-valued scalar specifying the maximum number of iterations without performance improvement the algorithm runs before stopping. Default is 70.

loss.unit
Loss function unit. A character-valued scalar indicating whether performance loss should be evaluated at the level of individual respondents (individuals) or geographic units (L2 units). Default is individuals.

loss.fun
Loss function. A character-valued scalar indicating whether prediction loss should be measured by the mean squared error (MSE) or the mean absolute error (MAE). Default is MSE.

lambda
Lasso penalty parameter. A numeric vector of non-negative values or a list of two numeric vectors of equal size, with the first vector containing the step sizes by which the penalty parameter should increase and the second vector containing the upper thresholds of the intervals to which the step sizes apply. The penalty parameter controls the shrinkage of the context-level variables in the lasso model. Default is list(c(0.1, 0.3, 1), c(1, 10, 10000)).

data
Data for cross-validation. A list of k data.frames, one for each fold to be used in k-fold cross-validation.

verbose
Verbose output. A logical argument indicating whether or not verbose output should be printed. Default is TRUE.

Examples

#not_yet

[Package autoMrP version 0.0.0.9000 Index]