run_svm {autoMrP}	R Documentation

Apply support vector machine classifier to MrP.

Description

run_svm is a wrapper function that applies the support vector machine classifier to data provided by the user, evaluates prediction performance, and chooses the best-performing model.

Usage

run_svm(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  kernel = "radial",
  loss.fun,
  gamma,
  cost,
  data,
  verbose
)

Arguments

y
Outcome variable. A character scalar containing the column name of the outcome variable in survey.

L1.x
Individual-level covariates. A character vector containing the column names of the individual-level variables in survey and census used to predict outcome y. Note that geographic unit is specified in argument L2.unit.

L2.x
Context-level covariates. A character vector containing the column names of the context-level variables in survey and census used to predict outcome y.

L2.unit
Geographic unit. A character scalar containing the column name of the geographic unit in survey and census at which outcomes should be aggregated.

L2.reg
Geographic region. A character scalar containing the column name of the geographic region in survey and census by which geographic units are grouped (L2.unit must be nested within L2.reg). Default is NULL.

kernel
SVM kernel. A character-valued scalar specifying the kernel to be used by SVM. The possible values are linear, polynomial, radial, and sigmoid. Default is radial.

loss.fun
SVM loss function. If NULL, then SVM uses the misclassification error to measure the loss of categorical predictions and the mean squared error to measure the loss of numeric predictions. Default is NULL.

gamma
SVM kernel parameter. A numeric vector whose values specify the gamma parameter in the SVM kernel. This parameter is needed for all kernel types except linear. Default is c(0.3, 0.5, 0.55, 0.6, 0.65, 0.7, 0.8, 0.9, 1, 2, 3, 4).

cost
SVM cost parameter. A numeric vector whose values specify the cost of constraints violation in SVM. Default is c(1, 10).

data
Data for cross-validation. A list of k data.frames, one for each fold to be used in k-fold cross-validation.

verbose
Verbose output. A logical argument indicating whether or not verbose output should be printed. Default is TRUE.

loss.unit
Loss function unit. A character-valued scalar indicating whether performance loss should be evaluated at the level of individual respondents (individuals) or geographic units (L2 units). Default is individuals.

Examples

#not_yet

[Package autoMrP version 0.0.0.9000 Index]