svm_classifier {autoMrP}	R Documentation

SVM classifier

Description

svm_classifier applies support vector machine classification to a data set.

Usage

svm_classifier(
  method,
  form,
  data,
  kernel,
  error.fun,
  probability,
  gamma.set,
  cost.set,
  sampling = "cross",
  cross,
  verbose = c(TRUE, FALSE)
)

Arguments

method
Function. A character string specifying the name of the function to be tuned.

form
Model formula. A two-sided linear formula describing the model to be fit, with the outcome on the LHS and the covariates separated by + operators on the RHS.

data
Data. A data.frame containing the cross-validation data used to train and evaluate the model.

kernel
Kernel for SVM. A character string specifying the kernel to be used for SVM. The possible types are linear, polynomial, radial, and sigmoid. Default is radial.

error.fun
function returning the error measure to be minimized. It takes two arguments: a vector of true values and a vector of predicted values. If NULL, the misclassification error is used for categorical predictions and the mean squared error for numeric predictions.

probability
Probability predictions. A logical argument indicating whether the model should allow for probability predictions

gamma.set
Gamma parameter for SVM. This parameter is needed for all kernels except linear.

cost.set
Cost parameter for SVM. This parameter specifies the cost of constraints violation.

sampling
Sampling scheme. A character string specifying the sampling scheme to be used. Possible values are cross, fix, and boot. Default is cross, which performs cross-validation.

verbose
Verbose output. A logical vector indicating whether or not verbose output should be printed.

Examples

#not_yet

[Package autoMrP version 0.0.0.9000 Index]